[hive] h.erase(it) no longer changes h.capacity()
Matt's reference implementation does a heuristic here that's
arguably worth doing: If we're erasing the last element from a
group, and the group has a successor, then deallocate the empty group
instead of returning it to the unused_groups list. This temporarily
reduces h.capacity(), meaning that we'll have to allocate again later;
but when we allocate again later, we'll allocate a bigger group,
so the data will end up being more contiguous as smaller groups
cycle out.

If we're removing the last element from the *tail* group (which is
generally speaking the biggest group), though, we must never
deallocate it, because that *would* churn the allocator needlessly.

Anyway, at least for now, never return any memory to the system
unless the user has actually requested it via shrink_to_fit
or trim_capacity.