[hive] MSVC compatibility fixes
MSVC doesn't allow anonymous struct types inside unions.
But Clang doesn't allow named struct types to be defined inside unions!
So, name the struct type *and* pull its definition outside of the union.
No intended functional change.

MSVC -W4 also warns about implicitly defaulted-as-deleted destructors;
so, explicitly delete these two destructors for clarity.

MSVC rejects a defaulted constructor for a type with a union involved
(they think it's defaulted-as-deleted). I think that's a bug, but it's
easy to fix. This constructor isn't trivial anyway, because of the
NSDMI on the `size_` member.