[inplace_vector] Give insert_range a stronger guarantee
If we have an unsized input range, we first append all the elements
to the end of the `inplace_vector`, and then rotate them into place.
But if we run out of capacity while appending, or if constructing an
element throws an exception, then the insert must fail.

If construction throws, then `ranges::uninitialized_copy_n` automatically
destroys the elements appended so far (strong exception guarantee).
If we run out of capacity, the old code would *not* destroy the elements
appended so far (basic exception guarantee) -- and in fact they'd be
left at the end, instead of in their "proper" inserted place. We can
destroy them without risking another exception, so let's do so
(thus giving the strong guarantee on out-of-capacity as well as on
construction-threw).